requirements.txt
Django==1.6
Markdown==2.5.1
Pillow==2.6.1
PyYAML==3.11
South==1.0.1
dj-database-url==0.3.0
dj-static==0.0.6
django-braces==1.4.0
django-filter==0.8
django-ipware==0.1.0
django-jsonfield==0.9.13
django-oauth-toolkit==0.7.2
django-phonenumber-field==0.6
django-rest-swagger==0.2.1
djangorestframework==2.4.4
djrill==1.2.0
gunicorn==19.1.1
jsonfield==1.0.0
oauthlib==0.7.1
phonenumbers==7.0.0
psycopg2==2.5.2
requests==2.4.3
simplejson==3.6.5
six==1.8.0
static3==0.5.1
wsgiref==0.1.2
